Fixed location of classes and removed null
